Final polish: split cli_config_apply.rs by scenario
Each of the four apply scenarios gets its own integration test binary:
- cli_config_apply_running: apply to a live runtime without restart
- cli_config_apply_local: backend update via local SQLite home
- cli_config_apply_graphql: backend update via live GraphQL endpoint
- cli_config_apply_e2e: tool services and scheduled tasks full round-trip

Each file is ~100-200 lines and carries only its own imports.
